requirements.txt
requests==2.23.0
selenium==3.141.0
mxnet==1.6.0
matplotlib==3.2.1
gluoncv==0.7.0
beautifulsoup4==4.9.1
Pillow==7.1.2
